// src/domains/csvRecords/types.ts
export interface CsvRecord {
  import_doc_num: string | null;
  IMPORT_LINE: string | null;
  cust_po: string | null;
  CO_PREFIX: string | null;
  PRODUCT_CODE: string | null;
  CUST_CODE: string | null;
  CUST_NAME: string | null;
  item: string | null;
  ser_num: string | null;
  import_price: string | null;
  qty_ordered: string | null;
  engine_model: string | null;
  engine_num: string | null;
  cust_num: string | null;
  end_user_cust_name: string | null;
  WORK_SCOPE: string | null;
  cert_num: string | null;
  order_date: string | null;
  part_rcvd_date: string | null;
  CSN_NUMBER: string | null;
  TSN_NUMBER: string | null;
}